removed a null check on $id which is assumed to be causing problems.
Oh, and it happens when we try to upload data as well so I have a
feeling it's something to do with the token...

On 18/10/2011, at 8:50 PM, Sam Watson wrote:

Hey guys,

Jad and I are having this problem with the app where it gets a response
of "error-3" from the server. '"error-3" is used for when the user
tries to either upload data outside the window or when the trial has
expired' to quote Enda. It's been happening since I think yesterday
afternoon. We haven't changed anything on our end that would affect it
so I don't know what it is. We get the error when we try to download
the data immediately after receiving a new token from the server
(logging in). Can we try and sort this out tonight please.

Sam